backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
youtube-transcript-api==0.6.1
pytube==15.0.0
transformers==4.30.1
torch==2.1.2
sentence-transformers==2.2.2
langchain==0.1.11
langchain-groq==0.1.1
langchain-community==0.0.28
python-dotenv==1.0.0
sacremoses==0.1.1
langchain-experimental==0.0.55
huggingface-hub==0.19.4